removing old test suite
